+++ CHANGELOG.md
@@ -36,6 +36,7 @@ and this project should strive to adhere to [Semantic Versioning](https://semver
 - Fix project allocation roll-up is not calculated correctly. Epics with children are ignored.
 - Fix Individual teams showed up as project graph.
 - Reword Epic tooltip on how allocations are used when ithas children nodes.
+- Fix multiple card renders on app load taking a long time.
 
 ## [v1.10.0] - 2026-02-21